Travis CI (#16)
* Travis?

* Added bazel hash

* Added missing #include, fixed a few sign comparison warnings.

* Added more warning to Xcode config

* Add missing include.

* Changed isnumber() to isdigit(), since isnumber() is Mac OS X specific.

* Removed #warning's

* Add missing include

* Removed dumb warning with Travis CI

+++ lemonscript/ExpressionScanner.h
@@ -49,7 +49,7 @@
     
     class lemonscript_expressions::ExpressionScanner {
         std::string toParse;
-        int readIndex = 0;
+        size_t readIndex = 0;
         
         bool got_eof = false; // true iff have seen EOF
         // (int rather than char to handle EOF)
